Combine datastore file and Delete unused files
