+++ README.md
@@ -4,5 +4,5 @@
 You can view my personal portfolio here: [Portfolio](https://xwang-sps-summer22.appspot.com/)
 </br>
 <br>
-You can contact me by: [LinkedIn](https://www.linkedin.com/in/xiaoyue-wang-611029232/); [Insagram](https://www.instagram.com/dp.moana.wang/): dp.moana.wang
+You can contact me by: [LinkedIn](https://www.linkedin.com/in/xiaoyue-wang-611029232/); [Insagram](https://www.instagram.com/dp.moana.wang/)
 </br>
