--- portfolio/src/main/java/com/google/sps/servlets/MyDatastore.java
@@ -1,35 +1,0 @@
-package com.google.sps.servlets;
-
-import com.google.cloud.datastore.Datastore;
-import com.google.cloud.datastore.DatastoreOptions;
-import com.google.cloud.datastore.Entity;
-import com.google.cloud.datastore.FullEntity;
-import com.google.cloud.datastore.KeyFactory;
-import java.io.IOException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/** Servlet responsible for creating new contact information. */
-@WebServlet("/store-data")
-public class MyDatastore extends HttpServlet {
-
-  @Override
-  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    // Sanitize user input to remove HTML tags and JavaScript.
-    String textValue = request.getParameter("text-input");
-    long timestamp = System.currentTimeMillis();
-
-    Datastore datastore = DatastoreOptions.getDefaultInstance().getService();
-    KeyFactory keyFactory = datastore.newKeyFactory().setKind("Contact");
-    FullEntity taskEntity =
-        Entity.newBuilder(keyFactory.newKey())
-            .set("contactinfo", textValue)
-            .set("timestamp", timestamp)
-            .build();
-    datastore.put(taskEntity);
-
-    response.sendRedirect("/index.html");
-  }
-}
