+++ portfolio/src/main/webapp/style.css
@@ -34,8 +34,8 @@ img{
 }
 
 textarea {
-    width: 300px;
-    height: 100px;
+    width: 400px;
+    height: 150px;
 }
 
 #contact-list {
