+++ walkthroughs/week-3-libraries/datastore/examples/todo-list/src/main/webapp/index.html
@@ -17,7 +17,7 @@
 
     <hr/>
 
-    <ul id="task-list"></ul>
+    <ul id="list-tasks"></ul>
 
   </body>
 </html>
